+++ internal/provider/resource_sdwan_system_ntp_feature.go
@@ -188,25 +188,25 @@ func (r *SystemNTPProfileParcelResource) Schema(ctx context.Context, req resourc
 							Optional:            true,
 						},
 						"hmac_sha2_value": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("HMAC-SHA2-256 (digest length = 256 bits, key length = [1-32] bytes), Attribute conditional on SD-WAN Manager version `26.1.0` or higher").String,
+							MarkdownDescription: helpers.NewAttributeDescription("HMAC-SHA2-256 (digest length = 256 bits, key length = [1-32] bytes), Attribute conditional on SD-WAN Manager version `26.1.1` or higher").String,
 							Optional:            true,
 							Validators: []validator.String{
 								stringvalidator.RegexMatches(regexp.MustCompile(`^\S{1,32}$`), ""),
 							},
 						},
 						"hmac_sha2_value_variable": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("Variable name, Attribute conditional on SD-WAN Manager version `26.1.0` or higher").String,
+							MarkdownDescription: helpers.NewAttributeDescription("Variable name, Attribute conditional on SD-WAN Manager version `26.1.1` or higher").String,
 							Optional:            true,
 						},
 						"cmac_aes128_value": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("CMAC-AES-128 (digest length = 128 bits, key length = [16 or 32] bytes), Attribute conditional on SD-WAN Manager version `26.1.0` or higher").String,
+							MarkdownDescription: helpers.NewAttributeDescription("CMAC-AES-128 (digest length = 128 bits, key length = [16 or 32] bytes), Attribute conditional on SD-WAN Manager version `26.1.1` or higher").String,
 							Optional:            true,
 							Validators: []validator.String{
 								stringvalidator.LengthBetween(16, 32),
 							},
 						},
 						"cmac_aes128_value_variable": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("Variable name, Attribute conditional on SD-WAN Manager version `26.1.0` or higher").String,
+							MarkdownDescription: helpers.NewAttributeDescription("Variable name, Attribute conditional on SD-WAN Manager version `26.1.1` or higher").String,
 							Optional:            true,
 						},
 					},